Vuln_scanner/requirements.txt
python-nmap
shodan


